requirements.txt
pip==20.2.3
setuptools==49.2.1
cryptography==3.4.7
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
asgiref==3.4.1
pycparser==2.20
requests==2.26.0
six==1.16.0
certifi==2021.5.30
urllib3==1.26.6
idna==3.2
django-crontab
pynacl